requirements.txt
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
docx2txt==0.8
google-generativeai>=0.3.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
pyinstaller>=6.0.0
